Refactor: move TakedownWatcher function to watch subcommand
