+++ internal/pkg/agent/cmd/watch_test.go
@@ -7,7 +7,10 @@ package cmd
 import (
 	"fmt"
 	"os"
+	"os/exec"
+	"path/filepath"
 	"runtime"
+	"strings"
 	"testing"
 	"time"
 
@@ -17,12 +20,14 @@ import (
 
 	"github.com/stretchr/testify/require"
 
+	"github.com/elastic/elastic-agent/internal/pkg/agent/application/filelock"
 	"github.com/elastic/elastic-agent/internal/pkg/agent/application/paths"
 	"github.com/elastic/elastic-agent/internal/pkg/agent/application/upgrade/details"
 	"github.com/elastic/elastic-agent/internal/pkg/agent/configuration"
 	"github.com/elastic/elastic-agent/internal/pkg/agent/errors"
 	"github.com/elastic/elastic-agent/internal/pkg/fleetapi"
 	"github.com/elastic/elastic-agent/internal/pkg/release"
+	"github.com/elastic/elastic-agent/pkg/core/logger"
 	"github.com/elastic/elastic-agent/pkg/core/logger/loggertest"
 
 	"github.com/elastic/elastic-agent/internal/pkg/agent/application/upgrade"
@@ -363,4 +368,234 @@ func Test_watchCmd(t *testing.T) {
 			}
 		})
 	}
+}
+
+func Test_takedownWatcher(t *testing.T) {
+
+	const applockerFileName = "mocklocker.lock"
+
+	testExecutablePath := filepath.Join("..", "application", "filelock", "testlocker", "testlocker")
+	if runtime.GOOS == "windows" {
+		testExecutablePath += ".exe"
+	}
+	testExecutableAbsolutePath, err := filepath.Abs(testExecutablePath)
+	require.NoError(t, err, "error calculating absolute test executable part")
+
+	require.FileExists(t, testExecutableAbsolutePath,
+		"testlocker binary not found.\n"+
+			"Check that:\n"+
+			"- test binaries have been built with mage build:testbinaries\n"+
+			"- the path of the executable is correct")
+
+	returnCmdPIDsFetcher := func(cmds ...*exec.Cmd) watcherPIDsFetcher {
+		return func() ([]int, error) {
+			pids := make([]int, 0, len(cmds))
+			for _, c := range cmds {
+				if c.Process != nil {
+					pids = append(pids, c.Process.Pid)
+				}
+			}
+
+			return pids, nil
+		}
+	}
+
+	// create a struct with a *exec.Cmd and a channel that will be closed when Wait() returns for the exec.Cmd
+	// this should keep the data race detector happy.
+	type testProcess struct {
+		cmd      *exec.Cmd
+		waitChan chan struct{}
+	}
+
+	type setupFunc func(t *testing.T, log *logger.Logger, workdir string) (watcherPIDsFetcher, []testProcess)
+	type assertFunc func(t *testing.T, workdir string, cmds []testProcess)
+
+	tests := []struct {
+		name               string
+		setup              setupFunc
+		wantErr            assert.ErrorAssertionFunc
+		assertPostTakedown assertFunc
+	}{
+		{
+			name: "no contention for watcher applocker",
+			setup: func(_ *testing.T, _ *logger.Logger, _ string) (watcherPIDsFetcher, []testProcess) {
+				// nothing to do here, always return and empty list of pids
+				return func() ([]int, error) {
+					return nil, nil
+				}, nil
+			},
+			wantErr: assert.NoError,
+			assertPostTakedown: func(t *testing.T, workdir string, _ []testProcess) {
+				// we should be able to lock, no problem
+				locker := filelock.NewAppLocker(workdir, applockerFileName)
+				lockError := locker.TryLock()
+				t.Cleanup(func() {
+					_ = locker.Unlock()
+				})
+
+				assert.NoError(t, lockError)
+
+			},
+		},
+		{
+			name: "contention with test binary listening to signals: test binary is terminated gracefully",
+			setup: func(t *testing.T, log *logger.Logger, workdir string) (watcherPIDsFetcher, []testProcess) {
+				cmd, testChan := createTestlockerCommand(t, log.Named("testlocker"), applockerFileName, testExecutableAbsolutePath, workdir, false)
+				require.NoError(t, err, "error starting testlocker binary")
+
+				// wait for test binary to acquire lock
+				require.EventuallyWithT(t, func(collect *assert.CollectT) {
+					assert.FileExists(collect, filepath.Join(workdir, applockerFileName), "watcher applocker should have been created by the test binary")
+				}, 10*time.Second, 100*time.Millisecond)
+				require.NotNil(t, cmd.Process, "process details for testlocker should not be nil")
+
+				t.Logf("started testlocker process with PID %d", cmd.Process.Pid)
+
+				return returnCmdPIDsFetcher(cmd), []testProcess{{cmd: cmd, waitChan: testChan}}
+			},
+			wantErr: assert.NoError,
+			assertPostTakedown: func(t *testing.T, workdir string, cmds []testProcess) {
+
+				assert.Len(t, cmds, 1)
+				testlockerProcess := cmds[0]
+				require.NotNil(t, testlockerProcess, "test locker process info should have a not nil cmd")
+
+				require.Eventually(t, func() bool {
+					running, checkErr := isProcessRunning(t, testlockerProcess.cmd)
+					if checkErr != nil {
+						t.Logf("error checking for testlocker process running: %s", checkErr.Error())
+						return false
+					}
+					return !running
+				}, 30*time.Second, 100*time.Millisecond, "test locker process should have exited")
+
+				<-testlockerProcess.waitChan
+
+				assert.True(t, testlockerProcess.cmd.ProcessState.Exited(), "test locker process should have terminated")
+				assert.Equal(t, 0, testlockerProcess.cmd.ProcessState.ExitCode(), "test locker process should have a successful exit status")
+
+				assert.FileExists(t, filepath.Join(workdir, applockerFileName))
+				testApplocker := filelock.NewAppLocker(workdir, applockerFileName)
+				testApplockerError := testApplocker.TryLock()
+				t.Cleanup(func() {
+					_ = testApplocker.Unlock()
+				})
+				assert.NoError(t, testApplockerError, "error locking applocker")
+			},
+		},
+		{
+			name: "contention with test binary not listening to signals: test binary is not terminated",
+			setup: func(t *testing.T, log *logger.Logger, workdir string) (watcherPIDsFetcher, []testProcess) {
+				cmd, waitChan := createTestlockerCommand(t, log.Named("testlocker"), applockerFileName, testExecutableAbsolutePath, workdir, true)
+				require.NoError(t, err, "error starting testlocker binary")
+
+				// wait for test binary to acquire lock
+				require.EventuallyWithT(t, func(collect *assert.CollectT) {
+					assert.FileExists(collect, filepath.Join(workdir, applockerFileName), "watcher applocker should have been created by the test binary")
+				}, 10*time.Second, 100*time.Millisecond)
+				require.NotNil(t, cmd.Process, "process details for testlocker should not be nil")
+
+				t.Logf("started testlocker process with PID %d", cmd.Process.Pid)
+
+				return returnCmdPIDsFetcher(cmd), []testProcess{{cmd: cmd, waitChan: waitChan}}
+			},
+			wantErr: assert.NoError,
+			assertPostTakedown: func(t *testing.T, workdir string, cmds []testProcess) {
+
+				assert.Len(t, cmds, 1)
+				testlockerProcess := cmds[0]
+				require.NotNil(t, testlockerProcess, "test locker process info should have exec.Cmd set")
+
+				// check that the process is still running for a time
+				assert.Never(t, func() bool {
+					running, checkErr := isProcessRunning(t, testlockerProcess.cmd)
+					if checkErr != nil {
+						t.Logf("error checking for testlocker process running: %s", checkErr.Error())
+						return false
+					}
+					return !running
+				}, 1*time.Second, 100*time.Millisecond, "test locker process should still be running for some time")
+
+				// Kill the process explicitly
+				err = testlockerProcess.cmd.Process.Kill()
+				assert.NoError(t, err, "error killing testlocker process")
+
+				<-testlockerProcess.waitChan
+
+				if assert.NotNil(t, testlockerProcess.cmd.ProcessState, "test locker process should have been terminated") {
+					assert.NotEqual(t, 0, testlockerProcess.cmd.ProcessState.ExitCode(), "test locker process should not return a successful exit code")
+				}
+			},
+		},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			workDir := t.TempDir()
+			log, obsLogs := loggertest.New(t.Name())
+			t.Cleanup(func() {
+				// however it ends, try to print out the logs of takedownWatcher
+				loggertest.PrintObservedLogs(obsLogs.All(), t.Log)
+			})
+			pidFetcher, processInfos := tc.setup(t, log, workDir)
+			tc.wantErr(t, takedownWatcher(t.Context(), log.Named("takedownWatcher"), pidFetcher))
+			if tc.assertPostTakedown != nil {
+				tc.assertPostTakedown(t, workDir, processInfos)
+			}
+		})
+	}
+}
+
+func createTestlockerCommand(t *testing.T, log *logger.Logger, applockerFileName string, testExecutablePath string, workdir string, ignoreSignals bool) (*exec.Cmd, chan struct{}) {
+
+	watchTerminated := make(chan struct{})
+
+	args := []string{"-lockfile", filepath.Join(workdir, applockerFileName)}
+	if ignoreSignals {
+		args = append(args, "-ignoresignals")
+	}
+
+	// use the same invoke as the one used to launch a watcher
+	watcherCmd, err := upgrade.StartWatcherCmd(log, func() *exec.Cmd {
+		cmd := upgrade.InvokeCmdWithArgs(testExecutablePath, args...)
+
+		cmd.Stdout = os.Stdout
+		cmd.Stderr = os.Stderr
+		return cmd
+	},
+		upgrade.WithWatcherPostWaitHook(func() {
+			close(watchTerminated)
+		}),
+	)
+
+	require.NoError(t, err, "error starting testlocker binary")
+	return watcherCmd, watchTerminated
+}
+
+func isProcessRunning(t *testing.T, cmd *exec.Cmd) (bool, error) {
+	if cmd.Process == nil {
+		return false, nil
+	}
+	t.Logf("checking if pid %d is still running", cmd.Process.Pid)
+	// search for the pid on the running processes
+	process, err := os.FindProcess(cmd.Process.Pid)
+	if err != nil {
+		t.Logf("error string: %q", err.Error())
+		if runtime.GOOS == "windows" && strings.Contains(err.Error(), "The parameter is incorrect") {
+			// in windows, noone can hear you scream
+			// invalid parameter means that the process object cannot be found
+			t.Logf("pid %d is not running because on windows we got an incorrect parameter error", cmd.Process.Pid)
+			return false, nil
+		}
+
+		t.Logf("error finding process: %T %v", err, err)
+		return false, err
+	}
+
+	if process == nil {
+		t.Logf("pid %d is not running because os.GetProcess returned a nil process", cmd.Process.Pid)
+		return false, nil
+	}
+
+	return isProcessLive(cmd.Process)
 }
